+++ config/medium/Hotel/code/CGCode.js
@@ -1,0 +1,15 @@
+const c = parseInt(readline());
+for (let i = 0; i < c; i++) {
+    const customer = readline();
+}
+const r = parseInt(readline());
+for (let i = 0; i < r; i++) {
+    const rule = readline();
+}
+for (let i = 0; i < c; i++) {
+
+    // Write an answer using console.log()
+    // To debug: console.error('Debug messages...');
+
+    console.log('customer floor');
+}